requirements.txt
uvicorn
fastapi
pydantic
python-dotenv
streamlit
scikit-learn